+++ inlong-tubemq/tubemq-server/src/main/java/org/apache/inlong/tubemq/server/master/metamanage/metastore/dao/mapper/AbstractMapper.java
@@ -23,6 +23,5 @@ public interface AbstractMapper {
 
     void close();
 
-    void loadConfig() throws LoadMetaException;
-
+    void loadConfig(StringBuilder strBuff) throws LoadMetaException;
 }
