+++ inlong-tubemq/tubemq-server/src/main/java/org/apache/inlong/tubemq/server/master/metamanage/metastore/dao/mapper/BrokerConfigMapper.java
@@ -24,23 +24,79 @@ import org.apache.inlong.tubemq.server.master.metamanage.metastore.dao.entity.Br
 
 public interface BrokerConfigMapper extends AbstractMapper {
 
-    boolean addBrokerConf(BrokerConfEntity memEntity,
+    /**
+     * Add a new broker configure info into store
+     *
+     * @param entity   need add record
+     * @param strBuff  the string buffer
+     * @param result   process result with old value
+     * @return  the process result
+     */
+    boolean addBrokerConf(BrokerConfEntity entity,
                           StringBuilder strBuff, ProcessResult result);
 
-    boolean updBrokerConf(BrokerConfEntity memEntity,
+    /**
+     * Update a broker configure info into store
+     *
+     * @param entity   need update record
+     * @param strBuff  the string buffer
+     * @param result   process result with old value
+     * @return  the process result
+     */
+    boolean updBrokerConf(BrokerConfEntity entity,
                           StringBuilder strBuff, ProcessResult result);
 
-    boolean delBrokerConf(int brokerId, ProcessResult result);
+    /**
+     * delete broker configure info from store
+     *
+     * @param brokerId  the broker id to be deleted
+     * @param strBuff   the string buffer
+     * @param result    the process result
+     * @return          whether success
+     */
+    boolean delBrokerConf(int brokerId, StringBuilder strBuff, ProcessResult result);
 
+    /**
+     * get broker configure info from store
+     *
+     * @param qryEntity    the query conditions
+     * @return result, only read
+     */
     Map<Integer, BrokerConfEntity> getBrokerConfInfo(BrokerConfEntity qryEntity);
 
+    /**
+     * get broker configure info from store
+     *
+     * @param brokerIdSet  need matched broker id set
+     * @param brokerIpSet  need matched broker ip set
+     * @param qryEntity    need matched properties
+     * @return result, only read
+     */
     Map<Integer, BrokerConfEntity> getBrokerConfInfo(Set<Integer> brokerIdSet,
                                                      Set<String> brokerIpSet,
                                                      BrokerConfEntity qryEntity);
 
+    /**
+     * get broker configure info from store
+     *
+     * @param brokerId  the broker id to be queried
+     * @return result, only read
+     */
     BrokerConfEntity getBrokerConfByBrokerId(int brokerId);
 
+    /**
+     * get broker configure info from store
+     *
+     * @param brokerIp   the broker ip to be queried
+     * @return result, only read
+     */
     BrokerConfEntity getBrokerConfByBrokerIp(String brokerIp);
 
+    /**
+     * get broker id set by region id set
+     *
+     * @param regionIdSet   the region id set
+     * @return  the regionId and brokerId set map
+     */
     Map<Integer, Set<Integer>> getBrokerIdByRegionId(Set<Integer> regionIdSet);
 }
