+++ inlong-tubemq/tubemq-server/src/main/java/org/apache/inlong/tubemq/server/master/metamanage/metastore/dao/mapper/TopicCtrlMapper.java
@@ -25,19 +25,61 @@ import org.apache.inlong.tubemq.server.master.metamanage.metastore.dao.entity.To
 
 public interface TopicCtrlMapper extends AbstractMapper {
 
+    /**
+     * Add a new topic control configure info into store
+     *
+     * @param entity   need add record
+     * @param strBuff  the string buffer
+     * @param result   process result with old value
+     * @return  the process result
+     */
     boolean addTopicCtrlConf(TopicCtrlEntity entity,
                              StringBuilder strBuff, ProcessResult result);
 
+    /**
+     * Update a topic control configure info into store
+     *
+     * @param entity   need update record
+     * @param strBuff  the string buffer
+     * @param result   process result with old value
+     * @return  the process result
+     */
     boolean updTopicCtrlConf(TopicCtrlEntity entity,
                              StringBuilder strBuff, ProcessResult result);
 
-    boolean delTopicCtrlConf(String topicName, ProcessResult result);
+    /**
+     * delete topic control configure info from store
+     *
+     * @param topicName  the topic name to be deleted
+     * @param strBuff  the string buffer
+     * @param result    the process result
+     * @return          whether success
+     */
+    boolean delTopicCtrlConf(String topicName, StringBuilder strBuff, ProcessResult result);
 
+    /**
+     * get topic control configure info from store
+     *
+     * @param topicName    the topic name to be queried
+     * @return result, only read
+     */
     TopicCtrlEntity getTopicCtrlConf(String topicName);
 
+    /**
+     * get topic control configure info from store
+     *
+     * @param qryEntity    the query conditions
+     * @return result, only read
+     */
     List<TopicCtrlEntity> getTopicCtrlConf(TopicCtrlEntity qryEntity);
 
+    /**
+     * get topic control configure info from store
+     *
+     * @param topicNameSet  need matched topic name set
+     * @param qryEntity     need matched properties
+     * @return result, only read
+     */
     Map<String, TopicCtrlEntity> getTopicCtrlConf(Set<String> topicNameSet,
                                                   TopicCtrlEntity qryEntity);
-
 }
