+++ inlong-tubemq/tubemq-server/src/main/java/org/apache/inlong/tubemq/server/master/metamanage/metastore/impl/bdbimpl/BdbBrokerConfigMapperImpl.java
@@ -22,37 +22,20 @@ import com.sleepycat.persist.EntityCursor;
 import com.sleepycat.persist.EntityStore;
 import com.sleepycat.persist.PrimaryIndex;
 import com.sleepycat.persist.StoreConfig;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 import org.apache.inlong.tubemq.corebase.rv.ProcessResult;
-import org.apache.inlong.tubemq.corebase.utils.ConcurrentHashSet;
 import org.apache.inlong.tubemq.server.common.exception.LoadMetaException;
 import org.apache.inlong.tubemq.server.master.bdbstore.bdbentitys.BdbBrokerConfEntity;
 import org.apache.inlong.tubemq.server.master.metamanage.DataOpErrCode;
 import org.apache.inlong.tubemq.server.master.metamanage.metastore.dao.entity.BrokerConfEntity;
-import org.apache.inlong.tubemq.server.master.metamanage.metastore.dao.mapper.BrokerConfigMapper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.apache.inlong.tubemq.server.master.metamanage.metastore.impl.AbsBrokerConfigMapperImpl;
 
-public class BdbBrokerConfigMapperImpl implements BrokerConfigMapper {
-
-    private static final Logger logger =
-            LoggerFactory.getLogger(BdbBrokerConfigMapperImpl.class);
-
+public class BdbBrokerConfigMapperImpl extends AbsBrokerConfigMapperImpl {
     // broker config store
     private EntityStore brokerConfStore;
     private final PrimaryIndex<Integer/* brokerId */, BdbBrokerConfEntity> brokerConfIndex;
-    private final ConcurrentHashMap<Integer/* brokerId */, BrokerConfEntity>
-            brokerConfCache = new ConcurrentHashMap<>();
-    private final ConcurrentHashMap<String/* brokerIP */, Integer/* brokerId */>
-            brokerIpIndexCache = new ConcurrentHashMap<>();
-    private final ConcurrentHashMap<Integer/* regionId */, ConcurrentHashSet<Integer>>
-            regionIndexCache = new ConcurrentHashMap<>();
 
     public BdbBrokerConfigMapperImpl(ReplicatedEnvironment repEnv, StoreConfig storeConfig) {
+        super();
         brokerConfStore = new EntityStore(repEnv,
                 TBDBStoreTables.BDB_BROKER_CONFIG_STORE_NAME, storeConfig);
         brokerConfIndex =
@@ -61,7 +44,7 @@ public class BdbBrokerConfigMapperImpl implements BrokerConfigMapper {
 
     @Override
     public void close() {
-        clearCacheData();
+        clearCachedData();
         if (brokerConfStore != null) {
             try {
                 brokerConfStore.close();
@@ -70,27 +53,27 @@ public class BdbBrokerConfigMapperImpl implements BrokerConfigMapper {
                 logger.error("[BDB Impl] close broker configure failure ", e);
             }
         }
+        logger.info("[BDB Impl] broker configure closed!");
     }
 
     @Override
-    public void loadConfig() throws LoadMetaException {
-        long count = 0L;
+    public void loadConfig(StringBuilder strBuff) throws LoadMetaException {
+        long totalCnt = 0L;
         EntityCursor<BdbBrokerConfEntity> cursor = null;
         logger.info("[BDB Impl] load broker configure start...");
+        // clear cached data
+        clearCachedData();
+        // load data from bdb
         try {
-            clearCacheData();
             cursor = brokerConfIndex.entities();
             for (BdbBrokerConfEntity bdbEntity : cursor) {
                 if (bdbEntity == null) {
                     logger.warn("[BDB Impl] found Null data while loading broker configure!");
                     continue;
                 }
-                BrokerConfEntity memEntity =
-                        new BrokerConfEntity(bdbEntity);
-                addOrUpdCacheRecord(memEntity);
-                count++;
+                addOrUpdCacheRecord(new BrokerConfEntity(bdbEntity));
+                totalCnt++;
             }
-            logger.info("[BDB Impl] total broker configure records are {}", count);
         } catch (Exception e) {
             logger.error("[BDB Impl] load broker configure failure ", e);
             throw new LoadMetaException(e.getMessage());
@@ -99,249 +82,15 @@ public class BdbBrokerConfigMapperImpl implements BrokerConfigMapper {
                 cursor.close();
             }
         }
-        logger.info("[BDB Impl] load broker configure successfully...");
+        logger.info(strBuff.append("[BDB Impl] loaded ").append(totalCnt)
+                .append(" broker configure successfully...").toString());
+        strBuff.delete(0, strBuff.length());
     }
 
-    @Override
-    public boolean addBrokerConf(BrokerConfEntity memEntity,
-                                 StringBuilder strBuff, ProcessResult result) {
-        BrokerConfEntity curEntity =
-                brokerConfCache.get(memEntity.getBrokerId());
-        if (curEntity != null) {
-            result.setFailResult(DataOpErrCode.DERR_EXISTED.getCode(),
-                    strBuff.append("The broker's brokerId ").append(memEntity.getBrokerId())
-                            .append(" has already exists, the value must be unique!")
-                            .toString());
-            strBuff.delete(0, strBuff.length());
-            return result.isSuccess();
-        }
-        Integer curBrokerId = brokerIpIndexCache.get(memEntity.getBrokerIp());
-        if (curBrokerId != null) {
-            result.setFailResult(DataOpErrCode.DERR_EXISTED.getCode(),
-                    strBuff.append("The broker's brokerIp ").append(memEntity.getBrokerIp())
-                            .append(" has already exists, the value must be unique!")
-                            .toString());
-            strBuff.delete(0, strBuff.length());
-            return result.isSuccess();
-        }
-        if (putBrokerConfig2Bdb(memEntity, strBuff, result)) {
-            addOrUpdCacheRecord(memEntity);
-        }
-        return result.isSuccess();
-    }
-
-    @Override
-    public boolean updBrokerConf(BrokerConfEntity memEntity,
-                                 StringBuilder strBuff, ProcessResult result) {
-        BrokerConfEntity curEntity =
-                brokerConfCache.get(memEntity.getBrokerId());
-        if (curEntity == null) {
-            result.setFailResult(DataOpErrCode.DERR_NOT_EXIST.getCode(),
-                    strBuff.append("The broker ").append(memEntity.getBrokerIp())
-                            .append("'s configure is not exists, please add record first!")
-                            .toString());
-            strBuff.delete(0, strBuff.length());
-            return result.isSuccess();
-        }
-        if (curEntity.equals(memEntity)) {
-            result.setFailResult(DataOpErrCode.DERR_UNCHANGED.getCode(),
-                    strBuff.append("The broker ").append(memEntity.getBrokerIp())
-                            .append("'s configure have not changed, please delete it first!")
-                            .toString());
-            strBuff.delete(0, strBuff.length());
-            return result.isSuccess();
-        }
-        if (putBrokerConfig2Bdb(memEntity, strBuff, result)) {
-            addOrUpdCacheRecord(memEntity);
-            result.setSuccResult(curEntity);
-        }
-        return result.isSuccess();
-    }
-
-    /**
-     * delete broker configure info from bdb store
-     *
-     * @param brokerId  the broker id to be deleted
-     * @param result    the process result
-     * @return          whether success
-     */
-    @Override
-    public boolean delBrokerConf(int brokerId, ProcessResult result) {
-        BrokerConfEntity curEntity =
-                brokerConfCache.get(brokerId);
-        if (curEntity == null) {
-            result.setSuccResult(null);
-            return result.isSuccess();
-        }
-        delBrokerConfigFromBdb(brokerId);
-        delCacheRecord(brokerId);
-        result.setSuccResult(curEntity);
-        return result.isSuccess();
-    }
-
-    /**
-     * get broker configure info from bdb store
-     *
-     * @param qryEntity    the query conditions
-     * @return result, only read
-     */
-    @Override
-    public Map<Integer, BrokerConfEntity> getBrokerConfInfo(BrokerConfEntity qryEntity) {
-        Map<Integer, BrokerConfEntity> retMap = new HashMap<>();
-        if (qryEntity == null) {
-            for (BrokerConfEntity entity : brokerConfCache.values()) {
-                retMap.put(entity.getBrokerId(), entity);
-            }
-        } else {
-            for (BrokerConfEntity entity : brokerConfCache.values()) {
-                if (entity != null && entity.isMatched(qryEntity)) {
-                    retMap.put(entity.getBrokerId(), entity);
-                }
-            }
-        }
-        return retMap;
-    }
-
-    /**
-     * get broker configure info from bdb store
-     *
-     * @param brokerIdSet  need matched broker id set
-     * @param brokerIpSet  need matched broker ip set
-     * @param qryEntity    need matched properties
-     * @return result, only read
-     */
-    @Override
-    public Map<Integer, BrokerConfEntity> getBrokerConfInfo(Set<Integer> brokerIdSet,
-                                                            Set<String> brokerIpSet,
-                                                            BrokerConfEntity qryEntity) {
-        Set<Integer> idHitSet = null;
-        Set<Integer> ipHitSet = null;
-        Set<Integer> totalMatchedSet = null;
-        Map<Integer, BrokerConfEntity> retMap = new HashMap<>();
-        // get records set by brokerIdSet
-        if (brokerIdSet != null && !brokerIdSet.isEmpty()) {
-            idHitSet = new HashSet<>();
-            BrokerConfEntity entity;
-            for (Integer brokerId : brokerIdSet) {
-                entity = brokerConfCache.get(brokerId);
-                if (entity != null) {
-                    idHitSet.add(brokerId);
-                }
-            }
-            if (idHitSet.isEmpty()) {
-                return retMap;
-            }
-        }
-        // get records set by brokerIpSet
-        if (brokerIpSet != null && !brokerIpSet.isEmpty()) {
-            ipHitSet = new HashSet<>();
-            for (String brokerIp : brokerIpSet) {
-                Integer brokerId = brokerIpIndexCache.get(brokerIp);
-                if (brokerId != null) {
-                    ipHitSet.add(brokerId);
-                }
-            }
-            if (ipHitSet.isEmpty()) {
-                return retMap;
-            }
-        }
-        // get intersection from brokerIdSet and brokerIpSet
-        if (idHitSet != null || ipHitSet != null) {
-            if (idHitSet == null) {
-                totalMatchedSet = new HashSet<>(ipHitSet);
-            } else {
-                if (ipHitSet == null) {
-                    totalMatchedSet = new HashSet<>(idHitSet);
-                } else {
-                    totalMatchedSet = new HashSet<>();
-                    for (Integer record : idHitSet) {
-                        if (ipHitSet.contains(record)) {
-                            totalMatchedSet.add(record);
-                        }
-                    }
-                }
-            }
-        }
-        // get broker configures
-        if (totalMatchedSet == null) {
-            for (BrokerConfEntity entity :  brokerConfCache.values()) {
-                if (entity == null
-                        || (qryEntity != null && !entity.isMatched(qryEntity))) {
-                    continue;
-                }
-                retMap.put(entity.getBrokerId(), entity);
-            }
-        } else {
-            for (Integer brokerId : totalMatchedSet) {
-                BrokerConfEntity entity = brokerConfCache.get(brokerId);
-                if (entity == null
-                        || (qryEntity != null && !entity.isMatched(qryEntity))) {
-                    continue;
-                }
-                retMap.put(entity.getBrokerId(), entity);
-            }
-        }
-        return retMap;
-    }
-
-    /**
-     * get broker configure info from bdb store
-     *
-     * @param brokerId  the broker id to be queried
-     * @return result, only read
-     */
-    @Override
-    public BrokerConfEntity getBrokerConfByBrokerId(int brokerId) {
-        return brokerConfCache.get(brokerId);
-    }
-
-    /**
-     * get broker configure info from bdb store
-     *
-     * @param brokerIp   the broker ip to be queried
-     * @return result, only read
-     */
-    @Override
-    public BrokerConfEntity getBrokerConfByBrokerIp(String brokerIp) {
-        Integer brokerId = brokerIpIndexCache.get(brokerIp);
-        if (brokerId == null) {
-            return null;
-        }
-        return brokerConfCache.get(brokerId);
-    }
-
-    @Override
-    public Map<Integer, Set<Integer>> getBrokerIdByRegionId(Set<Integer> regionIdSet) {
-        Set<Integer> qryKey = new HashSet<>();
-        Map<Integer, Set<Integer>> retInfo = new HashMap<>();
-        if (regionIdSet == null || regionIdSet.isEmpty()) {
-            qryKey.addAll(regionIndexCache.keySet());
-        } else {
-            qryKey.addAll(regionIdSet);
-        }
-        for (Integer regionId : qryKey) {
-            ConcurrentHashSet<Integer> brokerIdSet =
-                    regionIndexCache.get(regionId);
-            if (brokerIdSet == null || brokerIdSet.isEmpty()) {
-                continue;
-            }
-            retInfo.put(regionId, brokerIdSet);
-        }
-        return retInfo;
-    }
-
-    /**
-     * Put cluster setting info into bdb store
-     *
-     * @param memEntity need add record
-     * @param strBuff   the string buffer
-     * @param result process result with old value
-     * @return the process result
-     */
-    private boolean putBrokerConfig2Bdb(BrokerConfEntity memEntity,
-                                        StringBuilder strBuff, ProcessResult result) {
+    protected boolean putConfig2Persistent(BrokerConfEntity entity,
+                                           StringBuilder strBuff, ProcessResult result) {
         BdbBrokerConfEntity bdbEntity =
-                memEntity.buildBdbBrokerConfEntity();
+                entity.buildBdbBrokerConfEntity();
         try {
             brokerConfIndex.put(bdbEntity);
         } catch (Throwable e) {
@@ -356,7 +105,7 @@ public class BdbBrokerConfigMapperImpl implements BrokerConfigMapper {
         return result.isSuccess();
     }
 
-    private boolean delBrokerConfigFromBdb(int brokerId) {
+    protected boolean delConfigFromPersistent(int brokerId, StringBuilder strBuff) {
         try {
             brokerConfIndex.delete(brokerId);
         } catch (Throwable e) {
@@ -364,44 +113,5 @@ public class BdbBrokerConfigMapperImpl implements BrokerConfigMapper {
             return false;
         }
         return true;
-    }
-
-    private void delCacheRecord(int brokerId) {
-        BrokerConfEntity curEntity =
-                brokerConfCache.remove(brokerId);
-        if (curEntity == null) {
-            return;
-        }
-        brokerIpIndexCache.remove(curEntity.getBrokerIp());
-        ConcurrentHashSet<Integer> brokerIdSet =
-                regionIndexCache.get(curEntity.getRegionId());
-        if (brokerIdSet == null) {
-            return;
-        }
-        brokerIdSet.remove(brokerId);
-    }
-
-    private void addOrUpdCacheRecord(BrokerConfEntity entity) {
-        brokerConfCache.put(entity.getBrokerId(), entity);
-        // add brokerId info
-        Integer brokerId = brokerIpIndexCache.get(entity.getBrokerIp());
-        if (brokerId == null || brokerId != entity.getBrokerId()) {
-            brokerIpIndexCache.put(entity.getBrokerIp(), entity.getBrokerId());
-        }
-        ConcurrentHashSet<Integer> brokerIdSet = regionIndexCache.get(entity.getRegionId());
-        if (brokerIdSet == null) {
-            ConcurrentHashSet<Integer> tmpBrokerIdSet = new ConcurrentHashSet<>();
-            brokerIdSet = regionIndexCache.putIfAbsent(entity.getRegionId(), tmpBrokerIdSet);
-            if (brokerIdSet == null) {
-                brokerIdSet = tmpBrokerIdSet;
-            }
-        }
-        brokerIdSet.add(entity.getBrokerId());
-    }
-
-    private void clearCacheData() {
-        brokerIpIndexCache.clear();
-        regionIndexCache.clear();
-        brokerConfCache.clear();
     }
 }
